requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
aiomysql 
pymysql
pydantic[email]